Refactor corner enum
