+++ viewer/packages/tools/src/utilities/Corner.ts
@@ -1,5 +1,5 @@
 /*!
- * Copyright 2021 Cognite AS
+ * Copyright 2024 Cognite AS
  */
 
 export enum Corner {
